+++ silvera/lang/obj_processors.py
@@ -5,7 +5,7 @@ processors are used during parsing.
 from collections import deque, OrderedDict, defaultdict
 from silvera.core import (ServiceDecl, ConfigServerDecl, ServiceRegistryDecl,
                           TypedList, TypeDef, Deployable, Deployment,
-                          MessagePool, ProducerAnnotation, APIGateway)
+                          MessagePool, ProducerAnnotation, APIGateway, TypedSet)
 from silvera.exceptions import SilveraTypeError, SilveraLoadError
 from silvera.utils import available_port
 
@@ -212,7 +212,7 @@ def _resolve_fnc(module, service_decl, fnc):
     # Resolve function's return type
     ret_type = fnc.ret_type
     if not is_type_resolved(ret_type):
-        if isinstance(ret_type, TypedList):
+        if isinstance(ret_type, (TypedList, TypedSet)):
             _resolve_typed_list(service_decl, ret_type)
         else:
             try: